ajout toute autre fonctionalite no ajoute depuis deux semaine

+++ data/blog/home_index.php
@@ -60,24 +60,19 @@ https://templatemo.com/tm-596-electric-xtra
                 </svg>
                 <span class="logo-text"><?= replace_element_2(AsciiConverter::asciiToString($title_projet[0])) ?></span>
             </a>
-
-
             <ul class="nav-links" id="navLinks">
-
-
                 <?php
                 for ($i = 0; $i < count($title_projet_a); $i++) {
                 ?>
 
-                    <li><a href="#<?= $id_sha1_projet_a[$i] ?>"
-                            class="nav-link"><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?></a>
-                    </li>
+                <li><a href="#<?= $id_sha1_projet_a[$i] ?>"
+                        class="nav-link"><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?></a>
+                </li>
                 <?php
                 }
 
                 ?>
                 <li><a href="#Apropos" class="nav-link">À propos</a></li>
-
             </ul>
             <div class="menu-toggle" id="menuToggle">
                 <span></span>
@@ -86,11 +81,7 @@ https://templatemo.com/tm-596-electric-xtra
             </div>
         </div>
     </nav>
-
-
-
-
-    <!-- Hero Section -->
+        <!-- Hero Section -->
     <section class="hero" id="home">
         <div class="hero-content">
             <div class="text-rotator">
@@ -100,22 +91,30 @@ https://templatemo.com/tm-596-electric-xtra
                         <?= replace_element_2(AsciiConverter::asciiToString($title_projet[0])) ?></h1>
                     <p class="subtitle">
                         <?php
-                        $text = replace_element_2(AsciiConverter::asciiToString($description_projet[0]));
-                        echo mb_strlen($text, 'UTF-8') > 230
-                            ? mb_substr($text, 0, 230, 'UTF-8') . '...'
-                            : $text;
+ if ($description_projet_toggle[0] == "") {
+$description_projet_ = replace_element_1(AsciiConverter::asciiToString($description_projet[0])); 
+ }
+ else{
+$description_projet_ = replace_element_2(AsciiConverter::asciiToString($description_projet[0])); 
+ }
+
+                        if (strpos(replace_element_2(AsciiConverter::asciiToString($description_projet[0])), "ifram") !== false) {
+                            echo $description_projet_ ;                 
+                        
+                        } else {
+                            $text = $description_projet_;
+                            echo mb_strlen($text, 'UTF-8') > 230
+                                ? mb_substr($text, 0, 230, 'UTF-8') . '...'
+                                : $text;
+                        };
                         ?>
                     </p>
-
                 </div>
             </div>
         </div>
-
     </section>
-
     <?php
-
-    require_once "data/blog/carouselles/carouselle_2.php";
+        require_once "data/blog/carouselles/carouselle_2.php";
 
 
     ?>
@@ -131,15 +130,26 @@ https://templatemo.com/tm-596-electric-xtra
 
     <section class="about" id="about">
         <h2 class="section-title"><?= replace_element_2(AsciiConverter::asciiToString($title_projet[0])) ?></h2>
+
+
+        <?php
+
+     
+        ?>
+
+
         <div class="about-content">
             <div class="about-text">
 
-                <p><?= replace_element_2(AsciiConverter::asciiToString($description_projet[0])) ?></p>
+                <p><?= $description_projet_ ?></p>
             </div>
-            <div class="about-visual">
-                <div class="about-graphic"></div>
-            </div>
+
         </div>
+
+        <?php
+       
+
+        ?>
 
         <!-- Second row with reversed layout -->
 
@@ -155,37 +165,63 @@ https://templatemo.com/tm-596-electric-xtra
 
             if ($reste == 0) {
         ?>
-                <div id="<?= $id_sha1_projet_a[$i] ?>" class="about-content">
-                    <div class="about-visual">
-                        <div class="about-graphic-alt">
-                            <div class="hexagon"></div>
-                            <div class="hexagon"></div>
-                            <div class="hexagon"></div>
-                        </div>
-                    </div>
-                    <div class="about-text">
-                        <h2><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?> </h2>
-                        <p><?= replace_element_2(AsciiConverter::asciiToString($description_projet_a[$i])) ?></p>
-                    </div>
-
-
+        <div id="<?= $id_sha1_projet_a[$i] ?>" class="about-content">
+            <div class="about-visual">
+                <div class="about-graphic-alt">
+                    <div class="hexagon"></div>
+                    <div class="hexagon"></div>
+                    <div class="hexagon"></div>
                 </div>
+            </div>
+
             <?php
+
+
+
+                    if ($description_projet_toggle_a[$i] = "") {
+                    ?>
+
+            <div class="about-text">
+                <h2><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?> </h2>
+                <p><?= replace_element_2(AsciiConverter::asciiToString($description_projet_a[$i])) ?></p>
+            </div>
+
+            <?php
+                    } else {
+                    ?>
+
+            <div class="about-text">
+                <h2><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?> </h2>
+                <p><?= replace_element_1(AsciiConverter::asciiToString($description_projet_a[$i])) ?></p>
+            </div>
+            <?php
+                    }
+                    ?>
+
+
+        </div>
+        <?php
             } else {
-            ?>
 
-                <div id="<?= $id_sha1_projet_a[$i] ?>" class="about-content">
+                if ($description_projet_toggle_a[$i] = "") {
+                ?>
 
-                    <div class="about-text">
-                        <h2><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?> </h2>
-                        <p><?= replace_element_2(AsciiConverter::asciiToString($description_projet_a[$i])) ?></p>
-                    </div>
-                    <div class="about-visual">
-                        <div class="about-graphic"></div>
-                    </div>
+        <div class="about-text">
+            <h2><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?> </h2>
+            <p><?= replace_element_2(AsciiConverter::asciiToString($description_projet_a[$i])) ?></p>
+        </div>
 
-                </div>
-                <?php
+        <?php
+                } else {
+                ?>
+
+        <div class="about-text">
+            <h2><?= replace_element_2(AsciiConverter::asciiToString($title_projet_a[$i])) ?> </h2>
+            <p><?= replace_element_1(AsciiConverter::asciiToString($description_projet_a[$i])) ?></p>
+        </div>
+        <?php
+                }
+           
             }
 
 
@@ -206,12 +242,12 @@ https://templatemo.com/tm-596-electric-xtra
                     $img_user_b_ = '../img_dw/' . $img_user_b_;
                 ?>
 
-                    <div class="original">
-                        <img src="<?= $img_user_b_ ?>" alt="" srcset="">
+        <div class="original">
+            <img src="<?= $img_user_b_ ?>" alt="" srcset="">
 
-                    </div>
+        </div>
 
-            <?php
+        <?php
                 }
             }
 
@@ -227,7 +263,7 @@ https://templatemo.com/tm-596-electric-xtra
 
 
 
-            <a href="<?= $id_sha1_projet_a[$i] ?>" class="cta-button cta-primary">Voir page compléte</a>
+        <a href="<?= $id_sha1_projet_a[$i] ?>" class="cta-button cta-primary">Voir page compléte</a>
 
         <?php
 
@@ -238,14 +274,29 @@ https://templatemo.com/tm-596-electric-xtra
         ?>
     </section>
 
+
+    <style>
+        .wtop{
+            margin-top: 70px;
+        }
+        .section-title{
+            margin-top: 50px;
+            margin-bottom: 50px;
+
+        }
+         .cta-primary{
+            
+            margin-bottom: 150px;
+         }
+    </style>
     <section class="about" id="Apropos">
-        <h2 class="section-title">À propos</h2>
+        <h2 class="section-title wtop">À propos</h2>
         <div class="about-content">
 
 
 
 
-            <div class="about-text">
+            <div class="about-text" >
                 <?php
 
 
@@ -256,10 +307,10 @@ https://templatemo.com/tm-596-electric-xtra
                         $img_user_b_ = '../' . $img_user_b_;
                 ?>
 
-                        <div class="original">
-                            <img src="<?= $img_user_b_ ?>" alt="" srcset="">
+                <div class="original">
+                    <img src="<?= $img_user_b_ ?>" alt="" srcset="">
 
-                        </div>
+                </div>
 
                 <?php
                     }
@@ -307,7 +358,7 @@ https://templatemo.com/tm-596-electric-xtra
                 <?php
                 if ($id_sha1_parent_projet[0] != "") {
                 ?>
-                    <a href="<?= $id_sha1_parent_projet[0] ?>" class=""> ↰ Page parent</a>
+                <a href="<?= $id_sha1_parent_projet[0] ?>" class=""> ↰ Page parent</a>
 
 
                 <?php
@@ -324,7 +375,7 @@ https://templatemo.com/tm-596-electric-xtra
                 if (isset($_SESSION["password_projet"]) || isset($_SESSION["index"])) {
 
                 ?>
-                    <a href="../Class/session_destroy.php">off</a>
+                <a href="../Class/session_destroy.php">off</a>
                 <?php
                 }
                 ?>
@@ -341,47 +392,47 @@ https://templatemo.com/tm-596-electric-xtra
     <script src="../x.js"></script>
 
     <style>
-        .profil_user_image {
+    .profil_user_image {
 
-            width: 350px;
-            margin: auto;
+        width: 350px;
+        margin: auto;
 
-        }
+    }
 
-        .profil_user_image img {
-            width: 100%;
+    .profil_user_image img {
+        width: 100%;
 
 
-        }
+    }
 
-        .original {
-            width: 40%;
-            /* largeur du conteneur */
-            height: auto;
-            /* hauteur du conteneur s’adapte à l’image */
-            margin: auto;
-            display: flex;
-            justify-content: center;
-            /* centre horizontalement */
-            align-items: center;
-            /* centre verticalement si hauteur plus grande */
-            overflow: hidden;
-            /* cache le débordement si nécessaire */
+    .original {
+        width: 40%;
+        /* largeur du conteneur */
+        height: auto;
+        /* hauteur du conteneur s’adapte à l’image */
+        margin: auto;
+        display: flex;
+        justify-content: center;
+        /* centre horizontalement */
+        align-items: center;
+        /* centre verticalement si hauteur plus grande */
+        overflow: hidden;
+        /* cache le débordement si nécessaire */
 
-            margin-top: 45px;
-            margin-bottom: 45px;
+        margin-top: 45px;
+        margin-bottom: 45px;
 
-        }
+    }
 
-        .original img {
-            width: 100%;
-            /* prend toute la largeur du conteneur */
-            height: auto;
-            /* garde le ratio original */
-            display: block;
-            filter: brightness(0.6);
-            /* filtre sombre */
-        }
+    .original img {
+        width: 100%;
+        /* prend toute la largeur du conteneur */
+        height: auto;
+        /* garde le ratio original */
+        display: block;
+        filter: brightness(0.6);
+        /* filtre sombre */
+    }
     </style>
 </body>
